made enums modern, added new file block_data_config
