+++ python/python/ert/enkf/enums/enkf_obs_impl_type_enum.py
@@ -14,10 +14,9 @@
 #  See the GNU General Public License at <http://www.gnu.org/licenses/gpl.html>
 #  for more details.
 from cwrap import BaseCEnum
-from ert.enkf import ENKF_LIB
-
 
 class EnkfObservationImplementationType(BaseCEnum):
+    TYPE_NAME = "enkf_obs_impl_type"
     GEN_OBS = None
     SUMMARY_OBS = None
     BLOCK_OBS = None
@@ -25,8 +24,3 @@ class EnkfObservationImplementationType(BaseCEnum):
 EnkfObservationImplementationType.addEnum("GEN_OBS", 1)
 EnkfObservationImplementationType.addEnum("SUMMARY_OBS", 2)
 EnkfObservationImplementationType.addEnum("BLOCK_OBS", 3)
-
-EnkfObservationImplementationType.registerEnum(ENKF_LIB, "enkf_obs_impl_type")
-
-
-
