+++ python/python/ert/enkf/enums/enkf_run_enum.py
@@ -14,20 +14,16 @@
 #  See the GNU General Public License at <http://www.gnu.org/licenses/gpl.html>
 #  for more details.
 from cwrap import BaseCEnum
-from ert.enkf import ENKF_LIB
-
 
 class EnkfRunType(BaseCEnum):
+    TYPE_NAME = "enkf_run_mode_enum"
     ENKF_ASSIMILATION = None
     ENSEMBLE_EXPERIMENT = None
     SMOOTHER_UPDATED = None
     INIT_ONLY = None
-    
+
 
 EnkfRunType.addEnum("ENKF_ASSIMILATION" , 1)
 EnkfRunType.addEnum("ENSEMBLE_EXPERIMENT" , 2)
 EnkfRunType.addEnum("SMOOTHER_UPDATE" , 4)
 EnkfRunType.addEnum("INIT_ONLY" , 8)
-
-EnkfRunType.registerEnum( ENKF_LIB , "enkf_run_mode_enum")
-
