+++ python/python/ert/enkf/enums/ert_impl_type_enum.py
@@ -14,7 +14,6 @@
 #  See the GNU General Public License at <http://www.gnu.org/licenses/gpl.html>
 #  for more details.
 from cwrap import BaseCEnum
-from ert.enkf import ENKF_LIB
 
 
 class ErtImplType(BaseCEnum):
@@ -41,7 +40,3 @@ ErtImplType.addEnum("SUMMARY", 110)
 ErtImplType.addEnum("GEN_DATA", 113)
 ErtImplType.addEnum("SURFACE", 114)
 ErtImplType.addEnum("CONTAINER", 115)
-ErtImplType.registerEnum(ENKF_LIB, "ert_impl_type_enum")
-
-
-
