+++ python/python/ert/enkf/observations/__init__.py
@@ -1,4 +1,5 @@
 from .summary_observation import SummaryObservation
-from .block_observation import BlockObservation
 from .gen_observation import GenObservation
+from .block_data_config import BlockDataConfig
+from .block_observation import BlockObservation
 from .obs_vector import ObsVector
